+++ inst/doc/Bird_Parasite_CRF.R
@@ -27,7 +27,7 @@ plotMRF_hm(MRF_mod = MRF_fit, main = 'MRF (no covariates)',
                                             'Plasmodium', 'Microfilaria'))
 
 ## ------------------------------------------------------------------------
-net <- igraph::graph.adjacency(MRF_mod$graph, weighted = T, mode = "undirected")
+net <- igraph::graph.adjacency(MRF_fit$graph, weighted = T, mode = "undirected")
 igraph::plot.igraph(net, layout = igraph::layout.circle,
                     edge.width = abs(igraph::E(net)$weight),
                     edge.color = ifelse(igraph::E(net)$weight < 0,