+++ NEWS.md
@@ -1,5 +1,7 @@
 # MRFcov 1.0.37
 *   More detailed work-throughs of examples in the vignettes
+*   Fixed typo in Bird_Parasite_CRF vignette
+*   Fixed typo in cv_MRF_diag function
 
 # MRFcov 1.0.36
 *   Improved clarity of function definitions and descriptions